fixing warnings and some refactoring #12 #3

+++ src/netbuild/NBAlgorithms_Railway.cpp
@@ -991,7 +991,7 @@ NBRailwayTopologyAnalyzer::addBidiEdgesForStops(NBNetBuilder& nb) {
         delete t;
     }
     delete router;
-    return addBidiEdges.size();
+    return (int)addBidiEdges.size();
 }
